move byoda.util.secrets to byoda.secrets
