+++ byoda/datamodel/schema.py
@@ -21,7 +21,7 @@ from byoda.util import ServiceSignature
 from byoda.util import NetworkSignature
 from byoda.util import SignatureType
 
-from byoda.util.secrets import Secret, DataSecret
+from byoda.secrets import Secret, DataSecret
 
 from byoda.storage import FileStorage
